this pull request

+++ Build/LIBS/intel_osx_64/make_LIBS.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
-# use -I to force use of the Intel compiler
+arg=$1
+if [ "$arg" == "" ]; then
+  arg=all
+fi
+# use -I to force use of the gnu compiler
 OPTS="-I $*"
 source ../../../Source/scripts/setopts.sh $OPTS
 
@@ -7,7 +11,9 @@ DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 cd $DIR
 
 LIBDIR=`pwd`
-rm *.a
+if [[ "$arg" == "all" ]] || [[ "$arg" == "clean" ]]; then
+  rm *.a
+fi
 
 SRCDIR=$LIBDIR/../../../Source
 cd $SRCDIR
@@ -16,6 +22,7 @@ SRCDIR=`pwd`
 cd ../Build
 BUILDDIR=`pwd`
 
+if [[ "$arg" == "all" ]] || [[ "$arg" == "gd" ]]; then
 # GD
 echo
 echo "********** building GD" 
@@ -23,7 +30,9 @@ echo
 cd $SRCDIR/gd-2.0.15
 ./makelib.sh $OPTS
 cp libgd.a $LIBDIR/.
+fi
 
+if [[ "$arg" == "all" ]] || [[ "$arg" == "glui" ]]; then
 # GLUI
 echo
 echo "********** building glui"
@@ -31,8 +40,10 @@ echo
 cd $SRCDIR/glui_v2_1_beta
 ./makelib.sh $OPTS
 cp libglui.a $LIBDIR/.
+fi
 
 # GLUT
+if [[ "$arg" == "all" ]] || [[ "$arg" == "freeglut" ]]; then
 if [ "$GLUT" == "freeglut" ]; then
   cd $BUILDDIR/freeglut3.0.0/intel_osx_64
   ./make_freeglut.sh $OPTS 
@@ -44,7 +55,9 @@ else
     cp libglut.a $LIBDIR/.
   fi
 fi
+fi
 
+if [[ "$arg" == "all" ]] || [[ "$arg" == "jpeg" ]]; then
 # JPEG
 echo
 echo "********** building jpeg"
@@ -52,7 +65,9 @@ echo
 cd $SRCDIR/jpeg-9b
 ./makelib.sh $OPTS
 cp libjpeg.a $LIBDIR/.
+fi
 
+if [[ "$arg" == "all" ]] || [[ "$arg" == "png" ]]; then
 # PNG
 echo
 echo "********** building png"
@@ -60,7 +75,9 @@ echo
 cd $SRCDIR/png-1.6.21
 ./makelib.sh $OPTS
 cp libpng.a $LIBDIR/.
+fi
 
+if [[ "$arg" == "all" ]] || [[ "$arg" == "zlib" ]]; then
 # ZLIB
 echo
 echo "********** building zlib"
@@ -68,7 +85,9 @@ echo
 cd $SRCDIR/zlib128
 ./makelib.sh $OPTS
 cp libz.a $LIBDIR/.
+fi
 
+if [[ "$arg" == "all" ]] || [[ "$arg" == "lua" ]]; then
 if [ "$LUA" == "lua" ]; then
 
 # Lua # Lua interpreter
@@ -83,4 +102,4 @@ export TARGET=macosx
 ./makelib.sh $OPTS
 cp lpeg.so $LIBDIR/.
 fi
-
+fi
